starting working on pid for the turret, values not final
