+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Testing/Test.java
@@ -1,38 +1,25 @@
-package org.firstinspires.ftc.teamcode.Teleop;
+package org.firstinspires.ftc.teamcode.Testing;
 import com.bylazar.configurables.annotations.Configurable;
 import com.bylazar.telemetry.PanelsTelemetry;
 import com.bylazar.telemetry.TelemetryManager;
 import com.pedropathing.follower.Follower;
-import com.pedropathing.geometry.BezierLine;
 import com.pedropathing.geometry.Pose;
-import com.pedropathing.paths.HeadingInterpolator;
-import com.pedropathing.paths.Path;
-import com.pedropathing.paths.PathChain;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.teamcode.Mechanics.Robot;
 import org.firstinspires.ftc.teamcode.Mechanics.Shooter;
-import org.firstinspires.ftc.teamcode.Mechanics.Spind;
 import org.firstinspires.ftc.teamcode.Mechanics.Turret;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
-import java.util.function.Supplier;
-
+@Configurable
 @TeleOp
-public class ShooterData extends OpMode {
-    public static double power = 0;
+public class Test extends OpMode {
     private Follower follower;
     public static Pose startingPose = new Pose(72, 72, 0); //See ExampleAuto to understand how to use this
     private TelemetryManager telemetryM;
     private boolean slowMode = false;
     private double slowModeMultiplier = 0.5;
-    private boolean cooked = true;
-    private boolean julian = true;
-    private boolean coooooking = true;
-    private boolean coooooking2 = true;
-    private double spindPos = 0;
 
     @Override
     public void init() {
@@ -48,7 +35,7 @@ public class ShooterData extends OpMode {
         //The parameter controls whether the Follower should use break mode on the motors (using it is recommended).
         //In order to use float mode, add .useBrakeModeInTeleOp(true); to your Drivetrain Constants in Constant.java (for Mecanum)
         //If you don't pass anything in, it uses the default (false)
-        follower.startTeleopDrive();
+        follower.startTeleopDrive(true);
     }
 
     @Override
@@ -77,69 +64,28 @@ public class ShooterData extends OpMode {
             slowMode = !slowMode;
         }
 
-        //Optional way to change slow mode strength
-        if (gamepad1.xWasPressed()) {
-            slowModeMultiplier += 0.25;
-        }
+        telemetryM.debug("position", follower.getPose());
+        telemetryM.debug("velocity", follower.getVelocity());
 
-        //Optional way to change slow mode strength
-        if (gamepad2.right_bumper && cooked) {
-            Shooter.setHood(Shooter.getHood() + .01);
-            cooked = false;
+
+        if (gamepad2.x) {
+            Shooter.setPower(1);
         }
-        else if (!gamepad2.right_bumper){
-            cooked = true;
-        }
-        if (gamepad2.left_bumper && julian) {
-            Shooter.setHood(Shooter.getHood() - .01);
-            julian = false;
-        }
-        else if (!gamepad2.left_bumper){
-            julian = true;
+        if (gamepad2.b) {
+            Shooter.setPower(0);
         }
         if (gamepad2.right_trigger > .6) {
             Robot.transfer.setPosition(.9);
             try { Thread.sleep(100); } catch (Exception ignored) {}
             Robot.transfer.setPosition(.4);
         }
-        if (gamepad2.y && coooooking) {
-            spindPos += .5;
-//            int target = (int)Math.round((Constants.CPR / 6.0) * (x));
-//            Robot.spindexer.setTargetPosition(target);
-//            Robot.spindexer.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-//            Robot.spindexer.setPower(1);
-//            x += 1;
-            coooooking = false;
-        }
-        else if (!gamepad2.y)
-            coooooking = true;
-        if (gamepad2.a && coooooking2){
-//            int target = (int)Math.round((Constants.CPR / 6.0) * (x+1));
-//            Robot.spindexer.setTargetPosition(target);
-//            Robot.spindexer.setMode(DcMotorEx.RunMode.RUN_TO_POSITION);
-//            Robot.spindexer.setPower(1);
-//            x += 2;
-            spindPos += 1;
-            coooooking2 = false;
-        }
-        else if (!gamepad2.a)
-            coooooking2 = true;
-        if (gamepad2.x)
-            Shooter.setPower(power);
-        if (gamepad2.b)
-            Robot.flywheel.setPower(0);
-        if (gamepad2.left_trigger > .6)
-            Robot.flywheel.setPower(-.3);
 
-        telemetry.addData("distance", Math.sqrt(Math.pow(144 - follower.getPose().getX(),2) + Math.pow(144 - follower.getPose().getY(), 2)));
-        telemetry.addData("velocity", Shooter.getVel());
-        telemetry.addData("hood", Shooter.getHood());
-        Robot.intake.setPower(gamepad2.left_stick_y * Math.abs(gamepad2.left_stick_y));
-
-        Spind.spinTheDexer(spindPos);
 
         double t = Turret.faceGoal(follower.getPose().getX(), follower.getPose().getY(), follower.getHeading(), true, 0);
+        Shooter.autoShotHood(144 - follower.getPose().getX(), 144 - follower.getPose().getY());
 
+        telemetry.addData("hood", Shooter.autoShotHood(144 - follower.getPose().getX(), 144 - follower.getPose().getY()));
+        telemetry.addData("velocity", Shooter.getVel());
         telemetry.addData("x", follower.getPose().getX());
         telemetry.addData("y", follower.getPose().getY());
         telemetry.addData("theta", follower.getHeading());
