+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/Constants.java
@@ -14,7 +14,8 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 public class Constants {
-    public static double CPR = ((((1+(46D/17))) * (1+(46D/11))) * 28);
+    public static double CPR312 = ((((1+(46D/17))) * (1+(46D/11))) * 28);
+    public static double CPR1150 = ((1+(46D/11)) * 28);
     public static FollowerConstants followerConstants = new FollowerConstants()
             .mass(10.7)
             .forwardZeroPowerAcceleration(-59.57090583619754)
